SFX - Implemented Lotsa Sound Effects
I made a script that wraps an AudioSource for versatility, added it around, and created a SoundBank and corresponding enum values for that bank to allow a variety of sound effects to be standardized and used by the AudioSources.
SFX was implemented in various level element scripts.
In addition, very little post-processing was added to not demolish the listener's ears in a densely accessed area.

:llama: