DEV - Hotfixed Falling Error, Checkpoint Numbering
A previous approach to respawning the player wasn't working consistently so I used a collider to place them at the last reached checkpoint.
The checkpoints in the LevelManager weren't properly sorted in order and I didn't want to use a static ordering system, so I did it based on the numbers in the names for easy editor copy pasting. This approach stinks though, and will need to be improved.
:llama: